feat: remove the NTP syncing from server
