+++ meteor/server/api/systemTime/systemTime.ts
@@ -1,15 +1,6 @@
-import { Meteor } from 'meteor/meteor'
 const ntpClient: NtpClient = require('ntp-client')
 import { NtpClient } from '../../typings/ntp-client'
-import { systemTime, getCurrentTime } from '../../../lib/lib'
-import { setSystemStatus } from '../../systemStatus/systemStatus'
-import { logger } from '../../logging'
-import { TimeDiff, DiffTimeResult } from '../../../lib/api/peripheralDevice'
-import { env } from 'process'
-import { StatusCode } from '../../../lib/api/systemStatus'
-
-/** How often the system-time should be updated */
-const UPDATE_SYSTEM_TIME_INTERVAL = 3600 * 1000
+import { DiffTimeResult } from '../../../lib/api/peripheralDevice'
 
 /**
  * Send a number of calls to ntp-server, and calculate most-probable diff
@@ -17,13 +8,13 @@ const UPDATE_SYSTEM_TIME_INTERVAL = 3600 * 1000
  * https://stackoverflow.com/questions/1228089/how-does-the-network-time-protocol-work
  * @param config config object
  */
-async function determineDiffTimeInner(config: Config): Promise<DiffTimeResult> {
-	const maxSampleCount = config.maxSampleCount || 20
-	const minSampleCount = config.minSampleCount || 10
-	const maxAllowedDelay = config.maxAllowedDelay || 500
-	const maxTries = config.maxTries || 20
-	const host = config.host || ''
-	const port = config.port || 0
+async function determineDiffTimeInner(config?: Config): Promise<DiffTimeResult> {
+	const maxSampleCount = config?.maxSampleCount || 20
+	const minSampleCount = config?.minSampleCount || 10
+	const maxAllowedDelay = config?.maxAllowedDelay || 500
+	const maxTries = config?.maxTries || 20
+	const host = config?.host || '0.se.pool.ntp.org'
+	const port = config?.port || 123
 
 	return new Promise<Array<ServerTime>>((resolve, reject) => {
 		const results: Array<ServerTime> = []
@@ -72,13 +63,13 @@ interface ServerTime {
 	serverTime: number
 	responseTime: number
 }
-async function getServerTime(host?: string, port?: number, timeout?: number): Promise<ServerTime> {
+async function getServerTime(host: string, port: number, timeout?: number): Promise<ServerTime> {
 	return new Promise((resolve, reject) => {
 		ntpClient.ntpReplyTimeout = timeout || 500
 
 		const sentTime = Date.now()
 		try {
-			ntpClient.getNetworkTime(host || '0.se.pool.ntp.org', port || 123, (err: any, date: Date) => {
+			ntpClient.getNetworkTime(host, port, (err: any, date: Date) => {
 				if (err) {
 					reject(err)
 					return
@@ -125,90 +116,6 @@ interface Config {
 	port?: number
 }
 
-/**
- * Update the system time (this function is run at an interval()
- * @param retries
- */
-function updateServerTime(retries: number = 0) {
-	let ntpServerStr: string | undefined = process.env.NTP_SERVERS
-	if (!ntpServerStr) {
-		ntpServerStr = '0.se.pool.ntp.org,1.se.pool.ntp.org,2.se.pool.ntp.org'
-	}
-	const ntpServer = (ntpServerStr.split(',') || [])[0] || 'pool.ntp.org' // Just use the first one specified, for now
-	logger.info(`System time: Verifying, using ntp-server "${ntpServer}"...`)
-
-	determineDiffTimeInner({
-		host: ntpServer,
-		maxSampleCount: 20,
-		minSampleCount: 10,
-		maxAllowedDelay: 500,
-	})
-		.then((result) => {
-			// if result.stdDev is less than one frame-time, it should be okay:
-			if (result.stdDev < 1000 / 50) {
-				const message = `Diff is ${Math.round(result.mean)} ms (std. dev: ${
-					Math.floor(result.stdDev * 10) / 10
-				} ms)`
-				logger.info(`System time: ${message}`)
-
-				systemTime.hasBeenSet = true
-				systemTime.diff = result.mean
-				systemTime.stdDev = result.stdDev
-				setSystemStatus('systemTime', {
-					statusCode: systemTime.diff < 200 ? StatusCode.GOOD : StatusCode.WARNING_MAJOR,
-					messages: [message],
-				})
-			} else {
-				if (result.stdDev < systemTime.stdDev) {
-					systemTime.hasBeenSet = true
-					systemTime.diff = result.mean
-					systemTime.stdDev = result.stdDev
-				}
-				const message = `Unable to accuire NTP-time with good enough accuracy (standard deviation: ${
-					Math.floor(result.stdDev * 10) / 10
-				} ms)`
-				if (systemTime.stdDev < 200) {
-					setSystemStatus('systemTime', { statusCode: StatusCode.WARNING_MINOR, messages: [message] })
-				} else {
-					setSystemStatus('systemTime', { statusCode: StatusCode.WARNING_MINOR, messages: [message] })
-				}
-				Meteor.setTimeout(() => {
-					updateServerTime()
-				}, 20 * 1000)
-			}
-		})
-		.catch((err) => {
-			if (retries) {
-				Meteor.setTimeout(() => {
-					updateServerTime(retries - 1)
-				}, 1 * 1000)
-			} else {
-				logger.info('Unable to verify system time (' + (err.reason || err) + ')')
-				setSystemStatus('systemTime', {
-					statusCode: StatusCode.WARNING_MINOR,
-					messages: [`Error message: ${err.toString()}`],
-				})
-			}
-		})
-}
-Meteor.startup(() => {
-	if (!Meteor.isTest) {
-		setSystemStatus('systemTime', { statusCode: StatusCode.BAD, messages: ['Starting up...'] })
-		Meteor.setInterval(() => {
-			updateServerTime()
-		}, UPDATE_SYSTEM_TIME_INTERVAL)
-		updateServerTime(5)
-	} else {
-		systemTime.hasBeenSet = true
-		systemTime.diff = 0
-		systemTime.stdDev = 0
-		setSystemStatus('systemTime', {
-			statusCode: StatusCode.GOOD,
-			messages: [`NTP-time accuracy (standard deviation): 0 ms`],
-		})
-	}
-})
-
 // Example usage:
 // determineDiffTime({
 // 	maxSampleCount: 20,
@@ -219,20 +126,12 @@ Meteor.startup(() => {
 // 	logger.debug('result', result)
 // 	// if result.stdDev is less than one frame-time, we should be okay
 // })
-export async function determineDiffTime() {
+export async function determineDiffTime(config?: Partial<Config>) {
 	return determineDiffTimeInner({
 		maxSampleCount: 20,
 		minSampleCount: 10,
 		maxAllowedDelay: 500,
-	})
-}
 
-export function getTimeDiff(): TimeDiff {
-	return {
-		currentTime: getCurrentTime(),
-		systemRawTime: Date.now(),
-		diff: systemTime.diff,
-		stdDev: systemTime.stdDev,
-		good: systemTime.stdDev < 1000 / 50,
-	}
+		...config,
+	})
 }
